+++ store/gcworker/gc_worker.go
